--- Misc/Controllers.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6fa4ddeddfca4b9ca9b6dbb864686282
-timeCreated: 1594664857
