--- Misc/Controllers/AreaTriggerController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 52c9672823eb483d84d6a6bb6bb0688d
-timeCreated: 1595278827
