--- Misc/Controllers/FogAdjustmentController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 9a3ca04da03e4b2ca8434ee5cb7637af
-timeCreated: 1594664867
