--- Misc/Controllers/InteractTriggerController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 26fd66988c56481db9080505f286e4f6
-timeCreated: 1595297712
