--- Misc/Controllers/LerpedFollowerController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2d029997680745f1bc8849aa5128cb3f
-timeCreated: 1595146277
