--- Misc/Controllers/NetworkedTriggerController.cs
@@ -1,30 +1,0 @@
-﻿#if UNITY_EDITOR
-using UnityEngine;
-using VRC.Udon;
-using VRC.Udon.Common.Interfaces;
-
-namespace UdonToolkit {
-  [CustomName("Networked Trigger")]
-  [HelpMessage(
-    "This component waits for a \"Trigger\" custom event, e.g. from a UI Button, and calls a network event on all the provided behaviours. " +
-    "You can disable this object safely to save a bit of performance - it will still work as expected.")]
-  [HelpURL("https://github.com/orels1/UdonToolkit/wiki/Misc-Behaviours#networked-trigger")]
-  public class NetworkedTriggerController : UTController {
-    [SectionHeader("General")] [UdonPublic]
-    public NetworkEventTarget eventTarget;
-    
-    [ListView("Udon Events List")] [UdonPublic]
-    public UdonBehaviour[] udonTargets;
-
-    [ListView("Udon Events List")] [Popup(PopupAttribute.PopupSource.UdonBehaviour, "@udonTargets", true)] [UdonPublic]
-    public string[] udonEvents;
-
-    [Button("Trigger")]
-    public void Trigger() {
-      if (uB == null) return;
-      uB.SendCustomEvent("Trigger");
-    }
-
-  }
-}
-#endif
