--- Misc/Controllers/NetworkedTriggerController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8a7b48e5fcba4f4fbb0c236d2d62bb57
-timeCreated: 1595142221
