--- Misc/Controllers/ObjectBoundaryTriggerController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c5adae33831f4b899b677d7011268b15
-timeCreated: 1595290851
