--- Misc/Controllers/PlatformTriggerController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 524c6346a6914a709020109197edaa95
-timeCreated: 1597347799
