--- Misc/Controllers/SecretActionsController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f81384967771401f9633f48385d8f6e4
-timeCreated: 1595202878
