--- Misc/Controllers/ShaderFeederController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 995fe58d11e441dd9c2ea29fee0fc2b2
-timeCreated: 1597228462
