--- Misc/Controllers/SkyboxAdjustmentController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3b04e52a3ff54fc0bd49d297f3462980
-timeCreated: 1595188111
