--- Misc/Controllers/UIReadoutController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 5409863af230401aa5ee87db40b6bf9f
-timeCreated: 1595204286
