--- Misc/Controllers/UniversalTrackerController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: fa2caa22546c475f969791098ad7013f
-timeCreated: 1595119267
