+++ Misc/NetworkedTrigger.cs
@@ -15,9 +15,7 @@ namespace UdonToolkit {
       [ListView("Udon Events List")] [UTEditor]
       public UdonSharpBehaviour[] udonTargets;
       [ListView("Udon Events List")]
-      #if !COMPILER_UDONSHARP && UNITY_EDITOR
-      [Popup(PopupAttribute.PopupSource.UdonBehaviour, "@udonTargets", true)]
-      #endif
+      [Popup("behaviour", "@udonTargets", true)]
       [UTEditor]
       public string[] udonEvents;
