--- Misc/Program Assets/Area Trigger Udon C# Program Asset.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 96ab594a45b0c4a4397180d05a7c9726
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
